Struct secret::SecretValue

pub struct SecretValue {
    // some fields omitted
}

Methods

impl SecretValue

fn new(secret: &str) -> SecretValue

Create a SecretValue for the secret data passed in. This method will NOT create the actual secret in the backing Secret Service. (see SecretService.store() for that) The secret data is copied into non-pageable 'secure' memory.

fn get(&self) -> Option<String>

Get the secret data in the SecretValue. For now, this method only supports String values. Returns None, if the content type is npt text/plain

fn get_content_type(&self) -> String

Get the content type of the secret value, such as text/plain.

unsafe fn wrap(ptr: *mut SecretValueFFI) -> Self

Workaround method, do not use :-)